backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.7
asyncpg==0.29.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
pydantic-settings==2.1.0
chromadb==0.4.18
openai==1.3.6
google-generativeai==0.3.2
PyMuPDF==1.23.26
langchain==0.0.340
langchain-text-splitters==0.0.1
python-dotenv==1.0.0
supabase==2.0.2
postgrest==0.13.0
httpx==0.25.2